+++ src/main/kotlin/io/github/architpanigrahi/springbootdsl/dependency/DependencyDeclaration.kt
@@ -3,4 +3,5 @@ package io.github.architpanigrahi.springbootdsl.dependency
 data class DependencyDeclaration(
     val configuration: DependencyConfiguration,
     val notation: String,
+    val classifier: String? = null,
 )